initscripts: Add Test Manager
We need a way of automatically running test programs when the machine
boots. Install a luv-test-manager script and invoke it from runlevel 5.

luv-test-manager is a test manager, it is the file that drives the
entire test process. It searches  for test runner scripts in
/etc/luv-tests/ and executes any that it finds. Currently it only
handles execution of tests but later on it will also collect their
output and parse it into a standard format.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>