linux-yocto-efi-test: Enable kernel debugging options
There are a few kernel debug options that are not enabled in our kernel
config such as lockdep support, and RCU stall warning. Turn them on so
that we can catch these kinds of bugs at runtime.

There is a performance penalty to pay for having these features turned
on at runtime, but it's worth it because we favour correctness and ease
of debugging over performance. If our priorities change in future we can
always create a separate debug version of the kernel.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>