efivarfs: Ensure we build using target compiler
If we don't set ${CROSS_COMPILE} when building the efivarfs tests then
it will default to using the host's compiler, not the compiler for the
target. This results in the tests being unable to run on the target.

Use ${TARGET_PREFIX} to set ${CROSS_COMPILE} so that the efivarfs
Makefile will pickup the correct compiler.

Also, instead of building all the selftests limit the build to just the
efivarfs/ directory.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>