linux-yocto-efi-test: Enable all USB HCDs
The live image may be running from a USB flash drive, and because we
need to mount the image in order to access the kernel and initrd images
so that we can implement our crash solution, we need to make sure the
kernel includes drivers for the USB host controller.

Instead of worrying about the minimal number of drivers needed to work
with the majority hardware, just enable them all.

This results in a 2.59% increase in the kernel image size.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>