luv-live-image: Reduce grub timeout and add quiet kernel parameter
We don't need to wait for the default 10 seconds before booting the luv
kernel since the most common use case is people wanting to run the luv
tests, rather than interrupting the grub boot to inject some debugging
options.

Also add the "quiet" kernel parameter to reduce the noise of kernel
messages on the console.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>