luv-live-image: Create second partition for test results
Create a second partition to export the results of the LuvOS test
manager.

The idea behind using a different partition for the results is to avoid
any corruption on the boot partition.

The size of the partition is determined according to the following
assupmtions:

 - 500 test cases
 - 100 lines of raw results per test case
 -  20 lines of persed results per test case
 -  80 characters per line

Under such assumptions, 4.5MB of data space are required. As most USB
sticks have much greater capacity, the partition size is doubled to 8MB
to err on the safe side.

TODO: An arbitrary UUID is used for the results partitions. This UUID
must match with the UUID used by the test manager init script to mount
the partition. Thus, it is necessary to obtain the UUID by either
parsing such init script or modifying the UUID in the script

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Signed-off-by: Matt Fleming <matt.fleming@intel.com>