ovmf: Pass gcc version to build.sh
The build.sh build script in OvmfPkg allows us to optionally specify a
toolchain identifer, which we're currently not doing. By not passing a
toolchain ID directly build.sh invokes the build machine's compiler to
figure it out.

The important point in all of this is that the version identifier is
used in the ovmf build path, which means that if build machine's
compiler is upgraded/downgraded we can end up with multiple ovmf build
directories.

Pass the target's gcc version to build.sh instead.

Reported-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Signed-off-by: Matt Fleming <matt.fleming@intel.com>