fwts: Fix uefirttime failure due to using get_user() on kernel pointer
We've no need to turn a data structure on the stack into a pointer and
use get_user() on a kernel address, and in fact, doing so is likely to
return -EFAULT, because the argument isn't a valid user address.

This bug was introduced in commit c2268a9b23a9 ("fwts: Copied the
structure from userland locally in kernel space"), and causes a
uefirttime failure.

Cc: Pradeep Gaddam <pradeep.r.gaddam@intel.com>
Signed-off-by: Matt Fleming <matt.fleming@intel.com>