luv-live-image: Make luv-results the first partition
Windows will not mount and display more than one partition on a
removable disk.

Make the results partiion the first partition on the image so that users
are able to easily find the test results. Having this one partition
restriction isn't a problem otherwise, since there's little value in
auto-mounting the other partition (the EFI System Partition) anyway.

Of course, we're assuming firmware is smart enough to find the EFI
System Partition based on the partition table.

Fixes issue #23.

Reported-by: William Leara <williamleara@gmail.com>
Debugged-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Tested-by: Gayatri Kammela <gayatri.kammela@intel.com>
Signed-off-by: Matt Fleming <matt.fleming@intel.com>